added the reap function, fixed some signal issues

+++ signals.cpp
@@ -16,12 +16,17 @@ void ctrlZHandler(int sig_num) {
     }
     cout <<"Stopping the child!" << endl;
     kill(foregroundChildPID, SIGSTOP);
-    smashy.getJoblist()->addJob("my/name/is/maor/edri", foregroundChildPID, true); //usual suspect
+    smashy.getJoblist()->addJob("Anna zack", foregroundChildPID, true); //usual suspect !!
+    //smashy.getExternalCommandInFgPointer()->getCommand()
 }
 
 void ctrlCHandler(int sig_num) {
     auto& smashy = SmallShell::getInstance();
     pid_t foregroundChildPID = smashy.getForegroundCommandPID();
+    if (foregroundChildPID == 0){
+        cout << "No child to kill.. moron" << endl;
+        return;
+    }
     cout <<"Killing the child! sorry baby.." << endl;
     kill(foregroundChildPID, SIGTERM);
     cout <<"The child is dead" << endl;
@@ -37,7 +42,7 @@ void sigChildHandler(int sig_num){
     int status;
 
     cout << "a child has sent SIGCHLD" <<endl;
-    waitpid(-1, &status, WUNTRACED | WCONTINUED | WNOHANG);
+//    waitpid(-1, &status, WUNTRACED | WCONTINUED | WNOHANG);
 //    if(WIFCONTINUED(status)){
 //        printf("\n currentState = continued!\n");
 //    }
@@ -47,14 +52,14 @@ void sigChildHandler(int sig_num){
 //    if(WIFEXITED(status)){
 //        printf("\nterminattor\n");
 //    }
-    if(WIFSTOPPED(status)){
-        printf("\n currentState = stopped!\n");
-    }
-    cout << "went past checks";
-    if (smashy.getForegroundCommandPID() != 0){
-        return;
-    }
-    else{
-        wait(nullptr);
-    }
+//    if(WIFSTOPPED(status)){
+//        printf("\n currentState = stopped!\n");
+//    }
+//    cout << "went past checks";
+//    if (smashy.getForegroundCommandPID() != 0){
+//        return;
+//    }
+//    else{
+//        wait(nullptr);
+//    }
 }
